blueprint and MVC design are finished
