+++ auth/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+auth_obj = Blueprint('auth_obj', __name__)
+
+from . import auth
