+++ cart/views.py
@@ -1,13 +1,6 @@
 # # Define routes and views for this blueprint
 from flask import *
-# from flask import Blueprint
-
-# # Create a Blueprint instance
-# general = Blueprint('general', __name__)
-
-
 from . import booking, thankyou
-
 
 @booking.route("/booking")
 def booking():
