+++ sights/__init__.py
@@ -5,3 +5,4 @@ sights = Blueprint('sights', __name__, template_folder='templates/sights',
 
 # Import views after creating the blueprint to avoid circular imports
 from . import views
+from . import routes
